test: replace panic-checking tests with proper snapshot test
Removed tests that checked for panic messages and replaced with a proper
snapshot test that demonstrates graceful handling of malformed integrity
data. Uses normalizeBunSnapshot for consistent output.
